server/requirements.txt
fastapi
uvicorn
pydantic
pypdf
python-dotenv
langchain-community
langchain-text-splitters
langchain-pinecone
pinecone-client
python-multipart
langchain-groq 
langchain-huggingface
sentence-transformers
huggingface_hub